% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasster.r
\docType{package}
\name{fasster-package}
\alias{fasster}
\alias{fasster-package}
\title{fasster: Fast Additive Seasonal Switching with Trend and Exogenous Regressors}
\description{
Implementation of the FASSTER model for forecasting time series with multiple seasonalities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mitchelloharawild/fasster}
  \item Report bugs at \url{https://github.com/mitchelloharawild/fasster/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

}
\keyword{internal}
