% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_fasster.R
\name{FASSTER}
\alias{FASSTER}
\title{Fast Additive Switching of Seasonality, Trend and Exogenous Regressors}
\usage{
FASSTER(data, formula, include = NULL, ...)
}
\arguments{
\item{data}{A \code{ts}/\code{mts}, \code{tsibble} or \code{data.frame} that contains the variables in the model.}

\item{formula}{An object of class "formula" (refer to 'Formula' for usage)}

\item{include}{How many terms should be included to fit the model}

\item{...}{Not used}
}
\value{
Returns a mable containing the fitted FASSTER model.
}
\description{
Implements FASSTER
}
\details{
The fasster model extends commonly used state space models by introducing a switching component to the measurement equation.
This is implemented using a time-varying DLM with the switching behaviour encoded in the measurement matrix.
}
\section{Formula}{

\code{fasster} inherits the standard formula specification from \code{\link[stats]{lm}} for specifying exogenous regressors, including interactions and \code{\link[base]{I}()} functionality as described in \code{\link[stats]{formula}}.

Special DLM components can be specified using special functions defined below:
\itemize{
\item seas(s): Creates seasonal factors with seasonality s
\item trig(s): Creates seasonal fourier terms with seasonality s
\item poly(n): Creates a polynomial of order n (poly(1) creates a level, poly(2) creates a trend)
\item ARMA(ar, ma): Creates ARMA terms with coefficient vectors ar and ma
\item custom(dlm): Creates a custom dlm structure, using \code{\link[dlm]{dlm}}
}

The switching operator, \code{\%S\%} requires the switching factor variable on the LHS, and the model to switch over on the RHS (as built using the above components)
}

\examples{
tsibbledata::UKLungDeaths \%>\%
  FASSTER(mdeaths ~ fdeaths + poly(1) + trig(12))

}
