% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasster.r
\name{fitted.fasster}
\alias{fitted.fasster}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{fasster}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model fitted values is
    meaningful.}

\item{...}{other arguments.}
}
\value{
Fitted values extracted from the object \code{object}.
}
\description{
\code{fitted} is a generic function which extracts fitted values from
  objects returned by modeling functions.  \code{fitted.values} is an
  alias for it.

  All object classes which are returned by model fitting functions
  should provide a \code{fitted} method.  (Note that the generic is
  \code{fitted} and not \code{fitted.values}.)

  Methods can make use of \code{\link{napredict}} methods to compensate
  for the omission of missing values.  The default and \code{\link{nls}}
  methods do.
}
\references{
Chambers, J. M. and Hastie, T. J. (1992)
  \emph{Statistical Models in S}.
  Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{coefficients}}, \code{\link{glm}}, \code{\link{lm}},
  \code{\link{residuals}}.
}
