% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{ggfitted}
\alias{ggfitted}
\title{Time-series plot of fitted and observed values}
\usage{
ggfitted(object, ...)
}
\arguments{
\item{object}{A time-series model}

\item{...}{Additional arguments to be passed to \code{\link[forecast]{autoplot.mts}}}
}
\description{
Plots fitted and observed values with autoplot methods for \code{mts} using ggplot2.
}
\seealso{
\code{\link[forecast]{autoplot.mts}}
}
