% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate missing values using a model}
\usage{
interpolate(object, ...)
}
\arguments{
\item{object}{a mable}

\item{...}{}
}
\value{
A tsibble with missing values interpolated
}
\description{
Fills in the missing values from a model's response variable using the fitted values from the model.
}
