% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_tibble.fasster}
\alias{as_tibble.fasster}
\title{Coerce lists and matrices to data frames.}
\usage{
\method{as_tibble}{fasster}(x, ...)
}
\arguments{
\item{x}{A list. Each element of the list must have the same length.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
\code{\link[=as.data.frame]{as.data.frame()}} is effectively a thin wrapper around \code{data.frame},
and hence is rather slow (because it calls \code{\link[=data.frame]{data.frame()}} on each element
before \link{cbind}ing together). \code{as_tibble} is a new S3 generic
with more efficient methods for matrices and data frames.
}
\details{
This is an S3 generic. tibble includes methods for data frames (adds tbl_df
classes), tibbles (returns unchanged input), lists, matrices, and tables.
Other types are first coerced via \code{as.data.frame()} with
\code{stringsAsFactors = FALSE}.

\code{as_data_frame} and \code{as.tibble} are aliases.
}
\examples{
library(tsibble)
fasster(USAccDeaths ~ poly(1) + trig(12)) \%>\%
  as_tibble()

}
