% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as_tsibble.fasster}
\alias{as_tsibble.fasster}
\title{Coerce to a tsibble object}
\usage{
\method{as_tsibble}{fasster}(x, ...)
}
\arguments{
\item{x}{Other objects to be coerced to a tsibble (\code{tbl_ts}).}

\item{...}{Unquoted key variable(s) that define unique time indices. Only used
for data of class \code{tbl_df}, \code{data.frame} and \code{list} (i.e. long data form), if
a univariate time series (without an explicit key), simply leave it blank.
Ignored for other types of classes.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\section{Index}{


The time indices are no longer an attribute (for example, the \code{tsp} attribute
in a \code{ts} object), but preserved as the essential component of the tsibble. A
few index class, such as \code{Date}, \code{POSIXt}, and \code{difftime}, forms the basis of
the tsibble, with new additions \link{yearmth}, \link{yearqtr} and \link{year} representing
year-month, year-quarter and year objects respectively.

}

\section{Key}{


Key variable(s) together with the index uniquely identifies each record. And
key(s) also imposes the structure on a tsibble:
\itemize{
\item Univariate time series: none.
\item Multivariate time series: a single variable. For example, \code{data(pedestrian)}
uses the \code{Sensor} column as the key.
\item Grouped time series: multiple variables separated by commas (\code{,}). For example,
sales data over time can be split by product types (\code{product}) and geographical
locations (\code{geo}), and thus \code{product}, \code{geo} suggests the structure.
\item Hierarchical time series: nested keys using vertical bars (\code{|}). For example,
a natural hierarchy is defined on geographical regions: cities (\code{city}) are
nested in states (\code{state}). The expression of \code{city} | \code{state} is expected.
The general principle is that lower levels are nested in higher levels
(\code{lower} | \code{middle} | \code{higher}).
}

A combination of nested and grouped variables are also supported.

}

\section{Print options}{


The tsibble package fully utilises the \code{print} method from the tibble. Please
refer to \link[tibble:tibble-package]{tibble::tibble-package} to change display options.

}

