% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{FASSTER}
\alias{FASSTER}
\alias{fasster}
\title{Fast Additive Switching of Seasonality, Trend and Exogenous Regressors}
\usage{
FASSTER(formula, ...)
}
\arguments{
\item{formula}{An object of class "formula" (refer to 'Formula' for usage)}

\item{...}{Not used}

\item{data}{A \code{ts}/\code{mts}, \code{tsibble} or \code{data.frame} that contains the variables in the model.}

\item{include}{How many terms should be included to fit the model}
}
\value{
Returns a mable containing the fitted FASSTER model.
}
\description{
Implements FASSTER
}
\details{
The fasster model extends commonly used state space models by introducing a switching component to the measurement equation.
This is implemented using a time-varying DLM with the switching behaviour encoded in the measurement matrix.
}
\section{Formula}{

\code{fasster} inherits the standard formula specification from \code{\link[stats]{lm}} for specifying exogenous regressors, including interactions and \code{\link[base]{I}()} functionality as described in \code{\link[stats]{formula}}.

Special DLM components can be specified using special functions defined below:
\itemize{
\item seas(s): Creates seasonal factors with seasonality s
\item trig(s): Creates seasonal fourier terms with seasonality s
\item poly(n): Creates a polynomial of order n (poly(1) creates a level, poly(2) creates a trend)
\item ARMA(ar, ma): Creates ARMA terms with coefficient vectors ar and ma
\item custom(dlm): Creates a custom dlm structure, using \code{\link[dlm]{dlm}}
}

The switching operator, \code{\%S\%} requires the switching factor variable on the LHS, and the model to switch over on the RHS (as built using the above components)
}

\section{Heuristic}{

The model parameters are estimated using the following heuristic:
\enumerate{
\item Filter the data using the specified model with non-zero state variances
\item Obtain smoothed states \eqn{(\theta^{(s)}t=\theta_t|D_T)} to approximate correct behaviour
\item The initial state parameters taken from the first smoothed state: \eqn{m_0=E(\theta^{(s)}_0)}, \eqn{C_0=Var(\theta^{(s)}_0)}
\item Obtain state noise variances from the smoothed variance of \eqn{w_t}: \eqn{W=Var(w^{(s)}_t)=Var(\theta^{(s)}_t−G\theta^{(s)}_{t−1})}
   Obtain measurement noise variance from smoothed variance of \eqn{v_t}: \eqn{V=Var(v^{(s)}_t)=Var(y_t−F_t\theta^{(s)}_t)}
   \item Repair restricted state variances for seasonal factors and ARMA terms
}
}

\examples{
tsibbledata::UKLungDeaths \%>\%
  model(FASSTER(mdeaths ~ fdeaths + poly(1) + trig(12)))

}
