% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasster.r
\docType{package}
\name{fasster-package}
\alias{fasster-package}
\title{Fast Additive Switching of Seasonality, Trend and Exogenous Regressors}
\description{
Implementation of the FASSTER model for forecasting time series
with multiple seasonalities using switching states.
}
\details{
\tabular{ll}{ Package: \tab fasster\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/fasster}
  \item Report bugs at \url{https://github.com/tidyverts/fasster/issues}
}

}
\author{
Mitchell O'Hara-Wild

Maintainer: mail@mitchelloharawild.com
}
\keyword{package}
