% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasster.r
\name{fasster}
\alias{fasster}
\title{Fast Additive Seasonal Switching with Trend and Exogenous Regressors}
\usage{
fasster(data, model = y ~ intercept + trig(24) + trig(7 * 24) + xreg,
  lambda = NULL, heuristic = TRUE, ...)
}
\arguments{
\item{data}{A data.frame}

\item{model}{The fitted model}

\item{lambda}{Box-Cox transformation parameter. Ignored if NULL. Otherwise, data transformed before model is estimated. When lambda is specified, additive.only is set to TRUE.}

\item{heuristic}{Should the lm heuristic be used}

\item{...}{Not used}
}
\value{
Returns a fitted FASSTER model
}
\description{
Fast Additive Seasonal Switching with Trend and Exogenous Regressors
}
\examples{

}
