% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasster.r
\name{fasster}
\alias{fasster}
\title{Fast Additive Seasonal Switching with Trend and Exogenous Regressors}
\usage{
fasster(data, model = y ~ groupVar \%S\% (poly(1) + trig(24, 8)) + xreg,
  heuristic = c("filterSmooth", "lm", "saturated"), include = NULL,
  lambda = NULL, biasadj = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}

\item{model}{The fitted model}

\item{include}{How many terms should be included to fit the model}

\item{lambda}{Box-Cox transformation parameter. Ignored if NULL. Otherwise, data transformed before model is estimated. When lambda is specified, additive.only is set to TRUE.}

\item{...}{Not used}
}
\value{
Returns a fitted FASSTER model
}
\description{
Fast Additive Seasonal Switching with Trend and Exogenous Regressors
}
\examples{

}
