% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggfitted}
\alias{ggfitted}
\title{Time-series plot of fitted and observed values}
\usage{
ggfitted(object, ...)
}
\arguments{
\item{object}{A mable}

\item{...}{Additional arguments to be passed to \code{\link[fable]{autoplot.mable}}}
}
\description{
Plots fitted and observed values with autoplot methods for \code{tbl_ts} using ggplot2.
}
